Legger på typer for fakta-feilutbetaling
