+++ loosely-type-checked-files.json
@@ -135,7 +135,11 @@
   "packages/fakta-etablert-tilsyn/src/ui/mainReducer.ts",
   "packages/fakta-etablert-tilsyn/src/util/dateUtils.ts",
   "packages/fakta-etablert-tilsyn/src/util/periodUtils.ts",
+  "packages/fakta-feilutbetaling/src/components/FeilutbetalingInfoPanel.spec.tsx",
+  "packages/fakta-feilutbetaling/src/components/FeilutbetalingInfoPanel.tsx",
   "packages/fakta-feilutbetaling/src/components/FeilutbetalingPerioderForm.spec.tsx",
+  "packages/fakta-feilutbetaling/src/components/FeilutbetalingPerioderForm.tsx",
+  "packages/fakta-feilutbetaling/src/components/FeilutbetalingPerioderTable.spec.tsx",
   "packages/fakta-inntektsmelding/src/stories/Inntektsmelding.stories.tsx",
   "packages/fakta-inntektsmelding/src/ui/InntektsmeldingContainer.tsx",
   "packages/fakta-inntektsmelding/src/ui/components/arbeidsgiver-tekst/ArbeidsgiverTekst.tsx",
@@ -378,7 +382,7 @@
   "packages/sak-historikk/src/components/maler/HistorikkMalTypeTilbakekreving.spec.tsx",
   "packages/sak-historikk/src/components/maler/HistorikkMalTypeTilbakekreving.tsx",
   "packages/sak-historikk/src/components/maler/felles/Skjermlenke.tsx",
-  "packages/sak-historikk/src/components/maler/historikkMalType3.tsx",
+  "packages/sak-historikk/src/components/maler/historikkMalType4.tsx",
   "packages/sak-historikk/src/components/maler/historikkMalType5.tsx",
   "packages/sak-meldinger/src/MeldingerSakIndex.stories.tsx",
   "packages/sak-meldinger/src/components/Messages.tsx",
