--- packages/fakta-feilutbetaling/src/propTypes/feilutbetalingAksjonspunkterPropType.jsx
@@ -1,13 +1,0 @@
-// eslint-disable-next-line import/no-duplicates
-import PropTypes from 'prop-types';
-
-// eslint-disable-next-line import/no-duplicates
-import { kodeverkObjektPropType } from '@fpsak-frontend/prop-types';
-
-const feilutbetalingAksjonspunkterPropType = PropTypes.shape({
-  definisjon: kodeverkObjektPropType.isRequired,
-  status: kodeverkObjektPropType.isRequired,
-  kanLoses: PropTypes.bool.isRequired,
-});
-
-export default feilutbetalingAksjonspunkterPropType;
