--- packages/fakta-feilutbetaling/src/propTypes/feilutbetalingBehandlingPropType.jsx
@@ -1,8 +1,0 @@
-import PropTypes from 'prop-types';
-
-const feilutbetalingBehandlingPropType = PropTypes.shape({
-  id: PropTypes.number.isRequired,
-  versjon: PropTypes.number.isRequired,
-});
-
-export default feilutbetalingBehandlingPropType;
