+++ packages/utils/src/decodeHtmlEntity.ts
@@ -2,7 +2,7 @@
  * Decoder for decoding html entity navn og entity code
  * på sikkerhetshensyn '<(&lt; &#60;)' og '> (&gt; &#62;)' er ikke decodert
  */
-const decodeHtmlEntity = (str: string) => {
+const decodeHtmlEntity = (str: string | null | undefined) => {
   if (str === null || str === undefined) {
     return str;
   }
